// WADNR.Web/src/app/shared/generated/model/project-update-external-links-step.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectExternalLinkUpdateItem } from './project-external-link-update-item';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectUpdateExternalLinksStep { 
    ProjectUpdateBatchID?: number;
    ExternalLinks?: Array<ProjectExternalLinkUpdateItem> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectUpdateExternalLinksStepForm { 
    ProjectUpdateBatchID?: FormControl<number>;
    ExternalLinks?: FormControl<Array<ProjectExternalLinkUpdateItem>>;
}

export class ProjectUpdateExternalLinksStepFormControls { 
    public static ProjectUpdateBatchID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalLinks = (value: FormControlState<Array<ProjectExternalLinkUpdateItem>> | Array<ProjectExternalLinkUpdateItem> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectExternalLinkUpdateItem>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}