// WADNR.Web/src/app/shared/generated/model/project-update-external-links-step-request.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { ProjectExternalLinkUpdateItemRequest } from './project-external-link-update-item-request';


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectUpdateExternalLinksStepRequest { 
    ExternalLinks?: Array<ProjectExternalLinkUpdateItemRequest> | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectUpdateExternalLinksStepRequestForm { 
    ExternalLinks?: FormControl<Array<ProjectExternalLinkUpdateItemRequest>>;
}

export class ProjectUpdateExternalLinksStepRequestFormControls { 
    public static ExternalLinks = (value: FormControlState<Array<ProjectExternalLinkUpdateItemRequest>> | Array<ProjectExternalLinkUpdateItemRequest> = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<Array<ProjectExternalLinkUpdateItemRequest>>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}