// WADNR.Web/src/app/shared/generated/model/project-external-link-update-item.ts
/**
 * WADNR.API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


import { FormControl, FormControlOptions, FormControlState, Validators } from "@angular/forms";
export class ProjectExternalLinkUpdateItem { 
    ProjectExternalLinkUpdateID?: number;
    ProjectUpdateBatchID?: number;
    ExternalLinkLabel?: string | null;
    ExternalLinkUrl?: string | null;
    constructor(obj?: any) {
        Object.assign(this, obj);
    }
}

export interface ProjectExternalLinkUpdateItemForm { 
    ProjectExternalLinkUpdateID?: FormControl<number>;
    ProjectUpdateBatchID?: FormControl<number>;
    ExternalLinkLabel?: FormControl<string>;
    ExternalLinkUrl?: FormControl<string>;
}

export class ProjectExternalLinkUpdateItemFormControls { 
    public static ProjectExternalLinkUpdateID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ProjectUpdateBatchID = (value: FormControlState<number> | number = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<number>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalLinkLabel = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
    public static ExternalLinkUrl = (value: FormControlState<string> | string = undefined, formControlOptions?: FormControlOptions | null) => new FormControl<string>(
        value,
        formControlOptions ?? 
        {
            nonNullable: false,
            validators: 
            [
            ],
        }
    );
}